pyproject.toml
[tool.poetry]
name = "coursework2"
version = "0.1.0"
description = ""
authors = ["Aleksander <sanya.slepow2015@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.11.2"
black = "^24.8.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'
python_version = "3.12"

[tool.black]
# Максимальная длина строки
line-length = 119
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
